+++ botrunner/internal/game/result_table.go
@@ -7,8 +7,8 @@ type GameResultTableRow struct {
 	SessionTime    uint32  `yaml:"sessionTime"`
 	SessionTimeStr string  `yaml:"sessionTimeStr"`
 	HandsPlayed    uint32  `yaml:"handsPlayed"`
-	BuyIn          float32 `yaml:"buyIn"`
-	Profit         float32 `yaml:"profit"`
-	Stack          float32 `yaml:"stack"`
-	RakePaid       float32 `yaml:"rakePaid"`
+	BuyIn          float64 `yaml:"buyIn"`
+	Profit         float64 `yaml:"profit"`
+	Stack          float64 `yaml:"stack"`
+	RakePaid       float64 `yaml:"rakePaid"`
 }
